+++ codegen/package_test.go
@@ -69,8 +69,11 @@ func newPackageHelper(t *testing.T) *codegen.PackageHelper {
 	options := &codegen.PackageHelperOptions{
 		RelTargetGenDir: tmpDir,
 		CopyrightHeader: testCopyrightHeader,
-		GenCodePackage:  packageRoot + "/build/gen-code",
-		TraceKey:        "trace-key",
+		GenCodePackage: map[string]string{
+			".thrift": packageRoot + "/build/gen-code",
+			".proto":  packageRoot + "/build/gen-code",
+		},
+		TraceKey: "trace-key",
 	}
 
 	h, err := codegen.NewPackageHelper(